Build esy before publishing
