+++ src/installCommand/index.js
