+++ src/installCommand/logger.js
